Settings refactor (#1212)
