+++ .changeset/chilly-mugs-applaud.md
@@ -1,0 +1,5 @@
+---
+'@segment/analytics-core': patch
+---
+
+Refactor settings
