+++ .changeset/old-stingrays-tie.md
@@ -1,0 +1,5 @@
+---
+'@segment/analytics-core': patch
+---
+
+Update types / variable names
