+++ packages/browser/src/core/events/interfaces.ts
@@ -2,7 +2,7 @@ import {
   CoreOptions,
   CoreSegmentEvent,
   Callback,
-  Integrations,
+  IntegrationsOptions,
   Plan,
   TrackPlan,
   PlanEvent,
@@ -24,7 +24,7 @@ export type EventProperties = Record<string, any>
 export interface SegmentEvent extends CoreSegmentEvent {}
 
 export type {
-  Integrations,
+  IntegrationsOptions,
   Plan,
   TrackPlan,
   PlanEvent,
